Unify Create and Run Ingestion Pipeline Tasks
