+++ openmetadata-service/src/main/java/org/openmetadata/service/governance/workflows/WorkflowVariableHandler.java
@@ -3,6 +3,7 @@ package org.openmetadata.service.governance.workflows;
 import static org.openmetadata.service.governance.workflows.Workflow.FAILURE_VARIABLE;
 import static org.openmetadata.service.governance.workflows.Workflow.GLOBAL_NAMESPACE;
 
+import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
 import org.flowable.engine.delegate.DelegateExecution;
 import org.flowable.task.service.delegate.DelegateTask;
@@ -33,6 +34,11 @@ public class WorkflowVariableHandler {
     }
   }
 
+  public Object getNodeVariable(String varName) {
+    String namespace = getNodeNamespace();
+    return getNamespacedVariable(namespace, varName);
+  }
+
   public void setNamespacedVariable(String namespace, String varName, Object varValue) {
     String namespacedVarName = getNamespacedVariableName(namespace, varName);
     if (namespacedVarName != null) {
@@ -49,7 +55,8 @@ public class WorkflowVariableHandler {
 
   private String getNodeNamespace() {
     if (varScope instanceof DelegateExecution) {
-      return ((DelegateExecution) varScope).getParent().getCurrentActivityId();
+      return Optional.ofNullable(((DelegateExecution) varScope).getParent().getCurrentActivityId())
+          .orElseGet(() -> ((DelegateExecution) varScope).getCurrentActivityId());
     } else if (varScope instanceof DelegateTask) {
       return WorkflowHandler.getInstance()
           .getParentActivityId(((DelegateTask) varScope).getExecutionId());
