+++ openmetadata-spec/src/main/java/org/openmetadata/schema/governance/workflows/elements/WorkflowNodeDefinitionInterface.java
@@ -7,10 +7,8 @@ import java.util.List;
 import java.util.Map;
 import org.openmetadata.common.utils.CommonUtil;
 import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.CheckEntityAttributesTaskDefinition;
-import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.CreateIngestionPipelineTaskDefinition;
 import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.CreateAndRunIngestionPipelineTaskDefinition;
 import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.RunAppTaskDefinition;
-import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.RunIngestionPipelineTaskDefinition;
 import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.SetEntityCertificationTaskDefinition;
 import org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.SetGlossaryTermStatusTaskDefinition;
 import org.openmetadata.schema.governance.workflows.elements.nodes.endEvent.EndEventDefinition;
@@ -33,14 +31,8 @@ import org.openmetadata.schema.governance.workflows.elements.nodes.userTask.User
       name = "setGlossaryTermStatusTask"),
   @JsonSubTypes.Type(value = UserApprovalTaskDefinition.class, name = "userApprovalTask"),
   @JsonSubTypes.Type(
-      value = CreateIngestionPipelineTaskDefinition.class,
-      name = "createIngestionPipelineTask"),
-        @JsonSubTypes.Type(
-                value = RunIngestionPipelineTaskDefinition.class,
-                name = "runIngestionPipelineTask"),
-        @JsonSubTypes.Type(
-                value = CreateAndRunIngestionPipelineTaskDefinition.class,
-                name = "createAndRunIngestionPipelineTask"),
+      value = CreateAndRunIngestionPipelineTaskDefinition.class,
+      name = "createAndRunIngestionPipelineTask"),
   @JsonSubTypes.Type(value = RunAppTaskDefinition.class, name = "runAppTask"),
   @JsonSubTypes.Type(value = ParallelGatewayDefinition.class, name = "parallelGateway"),
 })
