+++ openmetadata-spec/src/main/resources/json/schema/governance/workflows/elements/nodeSubType.json
@@ -12,8 +12,6 @@
     "setGlossaryTermStatusTask",
     "endEvent",
     "startEvent",
-    "createIngestionPipelineTask",
-    "runIngestionPipelineTask",
     "createAndRunIngestionPipelineTask",
     "runAppTask",
     "parallelGateway"
