--- openmetadata-spec/src/main/resources/json/schema/governance/workflows/elements/nodes/automatedTask/runIngestionPipelineTask.json
@@ -1,87 +1,0 @@
-{
-  "$id": "https://open-metadata.org/schema/governance/workflows/elements/nodes/automatedTask/runIngestionPipelineTask.json",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "RunIngestionPipelineTaskDefinition",
-  "description": "Runs an Ingestion Pipeline based on its ID.",
-  "javaInterfaces": [
-    "org.openmetadata.schema.governance.workflows.elements.WorkflowNodeDefinitionInterface"
-  ],
-  "javaType": "org.openmetadata.schema.governance.workflows.elements.nodes.automatedTask.RunIngestionPipelineTaskDefinition",
-  "type": "object",
-  "definitions": {
-    "config": {
-      "type": "object",
-      "properties": {
-        "waitForCompletion": {
-          "title": "Wait for Completion",
-          "description": "Set if this step should wait until the Ingestion Pipeline finishes running",
-          "type": "boolean",
-          "default": true
-        },
-        "timeoutSeconds": {
-          "title": "Timeout Seconds",
-          "description": "Set the amount of seconds to wait before defining the Ingestion Pipeline has timed out.",
-          "type": "integer",
-          "default": 3600
-        }
-      },
-      "additionalProperties": false,
-      "required": ["waitForCompletion", "timeoutSeconds"]
-    }
-  },
-  "properties": {
-    "type": {
-      "type": "string",
-      "default": "automatedTask"
-    },
-    "subType": {
-      "type": "string",
-      "default": "runIngestionPipelineTask"
-    },
-    "name": {
-      "title": "Name",
-      "description": "Name that identifies this Node.",
-      "$ref": "../../../../../type/basic.json#/definitions/entityName"
-    },
-    "displayName": {
-      "title": "Display Name",
-      "description": "Display Name that identifies this Node.",
-      "type": "string"
-    },
-    "description": {
-      "title": "Description",
-      "description": "Description of the Node.",
-      "$ref": "../../../../../type/basic.json#/definitions/markdown"
-    },
-    "config": {
-      "$ref": "#/definitions/config"
-    },
-    "input": {
-      "type": "array",
-      "items": { "type":  "string" },
-      "default": ["ingestionPipelineId"],
-      "additionalItems": false,
-      "minItems": 1,
-      "maxItems": 1
-    },
-    "inputNamespaceMap": {
-      "type": "object",
-      "properties": {
-        "ingestionPipelineId": {
-          "type": "string",
-          "default": null
-        }
-      },
-      "additionalProperties": false,
-      "required": ["ingestionPipelineId"]
-    },
-    "branches": {
-      "type": "array",
-      "items": { "type":  "string" },
-      "default": ["success", "failure"],
-      "additionalItems": false,
-      "minItems": 2,
-      "maxItems": 2
-    }
-  }
-}
